add prompts

+++ install.ps1
@@ -7,7 +7,9 @@ function Setup-Neovim
 {
     Write-Host "Setup NeoVim"
     $filesFolder = $env:USERPROFILE + '\AppData\Local\nvim'
-    Remove-Item -Path "$filesFolder" -Recurse -Force
+    if (Test-Path -Path $filesFolder) {
+        Remove-Item -Path "$filesFolder" -Recurse -Force
+    }
     robocopy ".\" "$filesFolder" /E /MT /XD .git /XF .git* > $null
 }
